Fix out-of-bounds access
Prior to this change a corner case could cause an out-of-bounds access.

char *s = strdup("\r");
append_flags_sep_format(
  &s, APPEND_REMOVE_CR_LF_BEFORE | APPEND_REMOVE_CR_LF_AFTER, "", "");

In this case the removal of CR and LF will take place starting at 1 byte
before the start of the buffer, which results in an out-of-bounds read
and if CR or LF was found then an out-of-bounds write.

The parameter combination shown above is already covered in the tests
but the Visual Studio debugger doesn't have pointer checking that
detects the out-of-bounds access.